requirements.txt
tortoise-orm==0.24.0
fastapi==0.115.7
uvicorn[standard]==0.34.0
bcrypt==4.2.1
python-multipart==0.0.20